Refactor core parser
