+++ packages/core/src/external.rs
@@ -134,9 +134,11 @@ pub trait Builder<'t> {
 	/// Set the line_of_code offset of the error essages.
 	fn with_err_line(&mut self, line: usize) -> &mut dyn Builder<'t>;
 
-	// Set custom config. Used on derived Parser.
+	// WARNING: `Any` is not supported because trait object can't have generic methods
+
+	/// Set custom config. Used on derived Parser.
 	fn set(&mut self, key: &'static str, value: &'t str) -> &mut dyn Builder<'t>;
-	// Get custom config. Used on derived Parser.
+	/// Get custom config. Used on derived Parser.
 	fn get(&self, key: &'t str) -> Option<&'t str>;
 }
