+++ packages/core/src/semantics/transition/helper.rs
@@ -84,3 +84,73 @@ pub(super) mod get {
 		Event { name: event, guard }
 	}
 }
+
+/// analyze.rs helpers for transforming key for caches
+pub(super) mod transform_key {
+	use crate::semantics::*;
+
+	// WARNING: not performant because of using concatenated String as a key which cause filtering
+	impl From<&Event<'_>> for String {
+		fn from(event: &Event<'_>) -> Self {
+			format!("{}?{}", event.name.unwrap_or(""), event.guard.unwrap_or(""))
+		}
+	}
+
+	impl<'i> EventKey<'i> for &'i Option<String> {}
+	pub trait EventKey<'i>: Into<Option<&'i String>> {
+		fn has_trigger(self) -> bool {
+			self.into().filter(|e| is_empty(e.rsplit('?'))).is_some()
+		}
+		fn has_guard(self) -> bool {
+			self.into().filter(|e| is_empty(e.split('?'))).is_some()
+		}
+		fn get_guard(self) -> Option<&'i str> {
+			self.into().and_then(|e| none_empty(e.split('?')))
+		}
+		fn get_trigger(self) -> Option<&'i str> {
+			self.into().and_then(|e| none_empty(e.rsplit('?')))
+		}
+		fn guards_with_same_trigger(self, trigger: Option<&'i str>) -> Option<&'i str> {
+			self.into()
+				.filter(|e| none_empty(e.rsplit('?')) == trigger)
+				.and_then(|e| none_empty(e.split('?')))
+		}
+		fn triggers_with_same_guard(self, guard: Option<&'i str>) -> Option<&'i str> {
+			self.into()
+				.filter(|e| none_empty(e.split('?')) == guard)
+				.and_then(|e| none_empty(e.rsplit('?')))
+		}
+		fn as_expression(self) -> String {
+			self.into().map(String::as_str).as_expression()
+		}
+	}
+
+	impl<'o> Trigger<'o> for &'o Option<&'o str> {}
+	pub trait Trigger<'o>: Into<Option<&'o &'o str>> {
+		fn as_expression(self) -> String {
+			self.into()
+				.map(|s| {
+					format!(
+						" @ {trigger}{guard}",
+						trigger = none_empty(s.rsplit('?')).unwrap_or_default(),
+						guard = none_empty(s.split('?'))
+							.filter(|_| s.contains('?'))
+							.map(|g| format!("[{}]", g))
+							.unwrap_or_default(),
+					)
+				})
+				.unwrap_or_default()
+		}
+		fn as_key(self, guard: &str) -> Option<String> {
+			Some(format!("{}?{}", self.into().unwrap_or(&""), guard))
+		}
+	}
+
+	fn is_empty<'a>(split: impl Iterator<Item = &'a str>) -> bool {
+		none_empty(split).is_some()
+	}
+
+	fn none_empty<'a>(split: impl Iterator<Item = &'a str>) -> Option<&'a str> {
+		split.last().filter(|s| !s.is_empty())
+	}
+}
