+++ packages/transpiler/xstate/Cargo.toml
@@ -14,7 +14,7 @@ scdlang = { path = "../../core", version = "0.2.1" }
 serde_json = "1"
 serde = { version = "1", features = ["derive"] }
 serde_with = { version = "1", features = ["json"] }
-voca_rs = "1"
+voca_rs = "1" # helper to convert Scdlang naming convention into DavidKPiano naming convention
 
 [dev-dependencies]
 assert-json-diff = "1"
